+++ src/Symfony/Bundle/SecurityBundle/Tests/DependencyInjection/Fixtures/php/sodium_encoder.php
@@ -6,6 +6,8 @@ $container->loadFromExtension('security', [
     'encoders' => [
         'JMS\FooBundle\Entity\User7' => [
             'algorithm' => 'sodium',
+            'time_cost' => 8,
+            'memory_cost' => 128 * 1024,
         ],
     ],
 ]);